requirements.txt
Django==3.0.7
djangorestframework==3.12.2
django-phonenumber-field==4.0.0
wheel==0.34.2
social-auth-app-django==4.0.0
phonenumbers==8.12.6
Pillow==7.1.2
PyYAML==5.4.1 # used to generate schema into YAML-based OpenAPI format.
uritemplate==3.0.1 #used internally to get parameters in path for schema documentation
django-rest-swagger==2.2.0 #for Swagger UI